auth-service/requirements.txt
flask==2.0.1
sqlalchemy==1.4.23
pymysql==1.0.2
python-dotenv==0.19.0
werkzeug==2.0.1
email-validator==1.1.3 